requirements.txt
Flask>=2.0,<3.0
openai>=1.0,<2.0
python-dotenv>=0.19,<1.0
APScheduler>=3.6,<4.0
pytrends>=4.9.2,<5.0
pandas>=2.0,<3.0
gunicorn>=20.0,<21.0